refactor: 샘플 구조에 맞게 도메인 구조 재구성
- user/application/port/in, user/application/port/out, user/application/service 구조로 변경
- user/adapter/in/web, user/adapter/out/persistence 구조로 변경
- emotion/application/port/in, emotion/application/port/out, emotion/application/service 구조로 변경
- emotion/adapter/in/web, emotion/adapter/out/persistence 구조로 변경
- UserSelectedEmotionPersistencePort 파일 생성
- 패키지 선언 및 import 경로 수정 진행 중
