+++ src/main/java/com/melog/melog/user/adapter/in/web/UserController.java
@@ -1,9 +1,9 @@
-package com.melog.melog.user.adapter;
+package com.melog.melog.user.adapter.in.web;
 
 import com.melog.melog.common.model.request.UserCreateRequest;
 import com.melog.melog.common.model.request.UserUpdateRequest;
 import com.melog.melog.common.model.response.UserResponse;
-import com.melog.melog.user.application.UserUseCase;
+import com.melog.melog.user.application.port.in.UserUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
